+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Add new faraday middleware that automatically passes current `I18n.locale` to `lang` param
+
 ## [0.4.0] - 2023-02-09
 
 - Add a default read timeout for Faraday connection of 5 seconds, and open timeout of 1 second.